+++ lib/io/console.rb
@@ -1,0 +1,5 @@
+if RUBY_ENGINE == 'ruby'
+  require_relative 'console.so'
+else
+  require_relative 'console/ffi/console'
+end
