+++ lib/io/console/ffi/bsd_console.rb
