+++ lib/io/console/ffi/common.rb
