+++ lib/io/console/ffi/linux_console.rb
