+++ lib/io/console/ffi/native_console.rb
