+++ lib/io/console/ffi/stty_console.rb
