+++ lib/io/console/ffi/stub_console.rb
